+++ db/migrate/20160817003323_add_vendor_external_id_to_bloom_remit_txns.rb
@@ -1,0 +1,6 @@
+class AddVendorExternalIdToBloomRemitTxns < ActiveRecord::Migration
+  def change
+    add_column :bloom_remit_txns, :vendor_external_id, :string
+    add_index :bloom_remit_txns, :vendor_external_id
+  end
+end