+++ spec/concepts/bloom_remit/txns/operations/update_spec.rb
@@ -14,13 +14,16 @@ module BloomRemit
             secret: txn.secret,
             status: "paid",
             status_description: "My status description",
+            vendor_external_id: "somerefno",
           )
           txn = op.model
           expect(txn.status).to eq "paid"
           expect(txn.status_description).to eq "My status description"
+          expect(txn.vendor_external_id).to eq "somerefno"
           expect(OnUpdateJob).to have_enqueued_job(txn.id, {
             status: "paid",
             status_description: "My status description",
+            vendor_external_id: "somerefno",
           })
         end