+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Update `Txn#vendor_external_id` during callback from Bloom Remit
+
 ## [0.5.0] - 2016-08-09
 ### Added
 - Update `Txn#status_description` during callback from Bloom Remit